refactors methodology menu and pagination (#3586)
# Description and Motivation
fixes #3570 
This PR addresses the competing z-index values for the main navigation
menu and the sub-menus on the Methodology and Policy sections. It also
refactors these menus into dynamic HET components, for ease of usability
and scalability. It also refactors the Pagination buttons to achieve the
same purpose of dynamic reusability across the app.

## Has this been tested? How?
tests passing

## Screenshots (if appropriate)

## Types of changes
- Bug fix
- Refactor / chore

## New frontend preview link is below in the Netlify comment 😎